Handle invalid PBM files more gracefully.
A PBM file with a negative line width in the header could lead
to a crash due to array overrun in putwhitespan() (depending on
optimization and architecture).  Add check for negative values
for width and height - if yes, abort with error message.

Found and reported by Eric Sesterhenn <eric.sesterhenn@x41-dsec.de>

Signed-off-by: Gert Doering <gert@greenie.muc.de>