Enhance faxrunqd statistics.
Add "totals across all modems" block to existing statistics written
to LOG on exit/logrotate/SIGUSR2.

Add new type of statistics which is written in cyclic intervals in
CSV format to a dedicated file (one file per modem, or one file for
all).  These are not cumulative, so it's easier for humans to see
"here's a spike of errors", and also more convenient to plot - so
introduce a new %cyc_tty_statistics which basically duplicates the
existing %tty_statistics hash, but is cleared on every write-out.

Enable with "cyclic-statistics $time $filespec" in faxrunq.config.

Signed-off-by: Gert Doering <gd@medat.de>