BUSY is now also counted as "modem badness" (+0.1)
We used to count BUSY not as "a modem is broken" because it used
to be "a problem on the receiving end".  With all the new VoIP stuff,
and broken SIP PABXes, BUSY can nowadays also mean "this channel is
broken" (on the sending side).  So, take this into account.

Signed-off-by: Gert Doering <gd@medat.de>
(cherry picked from commit e858b6d14fda60614fb0e2419cd0d3f35a1cc44f)